requirements.txt
django-model-utils>=2.0
# Additional requirements go here
django-skosxl>=0.2
